перенест стили в ком-ты ком-ов ArticleForm, sidebar, CommentForm, UserInfo
